dataset-generation/requirements.txt
#Requires python version 3.10+

requests
httpx
aiohttp
beautifulsoup4
tqdm
sqlmodel
more-itertools
methodtools
natsort
python-debian
angr
pydemumble #Unvaiable for versions older than 3.10
typing-extensions
backports.strenum; python_version < "3.11"

#Python-magic-bin should include libmagic for windows, but is not available for linux
#Mac/Linux need to seperately install libmagic using apt, brew, etc.
python-magic-bin; platform_system  == "Windows"
python-magic; platform_system == "Linux" or platform_system == "Darwin"

dapper-python
dapper-python[dataset-generation]